+++ changelog.md
@@ -1,3 +1,10 @@
+# 1.0.2
+
+- Added an enable and disable button for the mod
+- Removed the option for the last level you played's music to continue
+- Fixed the mod so it doesnt break when playing levels you created
+- Added a togglable button in the pause menu to open the mod settings
+
 # 1.0.1
 
 - Added the option to disable the mod in practice mode